create overlay from operator, adds run/running workflow popovers
